Fix and adjustment from comments

+++ src/main/kotlin/com/zepben/evolve/cim/iec61970/base/wires/RotatingMachine.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Zeppelin Bend Pty Ltd
+ * Copyright 2024 Zeppelin Bend Pty Ltd
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -25,7 +25,7 @@ abstract class RotatingMachine @JvmOverloads constructor(mRID: String = "") : Re
 
     var ratedPowerFactor: Double? = null
     var ratedS: Double? = null
-    var ratedU: Double? = null
+    var ratedU: Int? = null
     var p: Double? = null
     var q: Double? = null
